requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
resend==0.6.0
python-dotenv==1.0.0
schedule==1.2.1
flask==3.0.2 